# Create database indices for unread items and deleted urls