@Minoru 2022-05-09
Weird, I can't reproduce the speedup. For me, this PR is consistently 1% faster than the current `master`, but that's it. This might be due to some settings I have in my config, or maybe I should experiment on a real disk instead of tmpfs. I'll try different configurations later and report back.

What I *can* reproduce is the size increase :) 4.4% for my database. It's okay as long as we get a comparable speed increase, I think.