@Minoru 2022-05-09
Ideas from IRC discussion with noctux and Evil_Bob:

- [ ] try this on a real disk, not tmpfs
- [ ] try applying the patch to the latest release, not master
    - noctux's measurements above weren't even done with this PR; the indices were added manually. Could our DB versioning be the culprit? I wonder