@Minoru 2022-05-11
Nope, can't reproduce. Things I tried:
- running the tests on HDD rather than tmpfs (with plenty of free RAM for the page cache, and one warm-up run before each 10-run test)
- cherry-picking this PR onto r2.27 tag (i.e. applying it to the latest release rather than `master`)
- adding the indices manually (actually, I ran the same binary on two different databases, one "stock", the other one already converted by an earlier invocation of the code from this PR)
- running with default settings (`--config-file=/dev/null`)

Each test run shows that this PR is 1% faster than the baseline, which is slightly comforting :)

I start to wonder if my data is somehow non-representative. @noctux, does your cache.db contain anything private? :) Would it be possible to share it with me (privately, under a promise that I won't pass it further and will destroy my copy once the mystery here is solved)?

Also, if you have a minute, can you please re-run your tests with `--config-file=/dev/null` to make sure that your config doesn't affect this? (It's not very likely that both you and Evil_Bob have a setting that I don't, but hey, it's easy enough to check)/