@Minoru 2022-05-28
Notes from a couple discussions we had with Evil_Bob and noctux on IRC:

- I made a subset of my cache.db, and both Evil_Bob and noctux **don't** see the expected speedup on it too! So my data is to blame for my (lack of) results above
- Evil_Bob dumped my subset cache.db, created a new one, and still couldn't observe the promised speedup: it went from 17 seconds on my file to 14 on his, but there is no difference between indexed and non-indexed cases
- my subset cache.db has a page size of 1024 bytes, whereas newly created databases have 4096 (I found that out using `.dbinfo` in `sqlite3` shell). If I take my subset cache.db, open sqlite3 shell, type `pragma page_size=4096; vacuum;`, then `echo qy | newsboat` goes from 20 seconds to 16.8 (on tmpfs). Still no difference between indiced and non-indiced cases though. The database with 4K pages is 0.9% smaller
- Evil_Bob wrote [a script](https://github.com/newsboat/newsboat/files/8791862/newsboat_create_feeds.sh.gz) that generates a bunch of feeds and a urls file that includes them. With this setup, I **can** reproduce the speedup promised by this PR: 15s went down to 3.6s (12s to 3.18s for Evil_Bob)

I want to get to the bottom of this, because I feel uneasy merging something that doesn't quite behave the way we expect it to. OTOH we haven't seen a case where this PR is *slower* than current master, so maybe I'll just cave in and merge it even if I don't understand the behaviour.

I think the next step is to look at what SQL queries should use these indices, and compare the output of SQL `ANALYZE` between my subset DB and the one generated by the script.